Add files via upload
Link prediction finds missing links or predicts the likelihood of future links. The problem of link prediction is a very well-studied problem and has applications in many different areas, some of which can be commonly seen in recommendation engines, such as new connections (suggested friends/followers) on social networks or related products on shopping sites.